Improves site accessibility and UI enhancements
Refactors the footer to use tooltips for icon links, enhancing discoverability and accessibility.
Updates the header and mobile menu to clarify the GitHub link as a "Repository."

Removes theme sanitization logic as it is no longer needed with the updated theme handling.

Additionally, improves the cookie consent banner by adding an info icon and enhancing link accessibility.

	deleted:    .dummy.ts
	modified:   .github/CONTRIBUTING.md
	modified:   app/layout.tsx
	modified:   components/lightModeTip.tsx
	modified:   components/main/footer.tsx
	modified:   components/main/header.tsx
	modified:   components/modules/constructStatusBadge.tsx
	modified:   components/modules/mobileMenu.tsx
	modified:   components/modules/modes.tsx
	deleted:    components/modules/themeSanitizer.ts
	modified:   components/scripts/cookieConsent.tsx
	new file:   hooks/useForceUnlockScroll.ts
	deleted:    hooks/useThemeNormalize.ts
	modified:   .github/CONTRIBUTING.md